+++ src/fsd_entities/auth/model/queries/usePostSignUp.ts
@@ -7,21 +7,21 @@ import toast from 'react-hot-toast';
 
 import { parseErrorMessage } from '@/fsd_shared';
 
-import { submitAdmissionsApplication } from '../../api/post';
+import { signup } from '../../api/post';
 
-export const usePostAdmission = () => {
+export const usePostSignUp = () => {
   const router = useRouter();
 
   return useMutation({
-    mutationFn: submitAdmissionsApplication,
+    mutationFn: signup,
     onSuccess: () => {
-      toast.success('가입 신청서 제출이 완료되었습니다!');
+      toast.success('회원가입이 완료되었습니다!');
       setTimeout(() => {
-        router.push('/auth/authorization');
+        router.push('/auth/signin');
       }, 500);
     },
-    onError: (error: any) => {
-      toast.error(parseErrorMessage(error, '가입 신청서 제출 실패: '));
+    onError: (error: Error.ApiErrorResponse) => {
+      toast.error(parseErrorMessage(error, '회원가입 중 오류가 발생했습니다.'));
     },
   });
 };
